refactor: game util static으로 추출
