+++ src/racingGameManager.js
@@ -1,14 +1,33 @@
 import Car from './car.js';
 import generateRandomNumber from './game/utils/generateRandomNumber.js';
 import isMovable from './game/utils/isMovable.js';
+import hasDuplicate from './utils/hasDuplicate.js';
+import isValidCarName from './game/utils/isValidCarName.js';
 import toCarNameList from './game/utils/toCarNameList.js';
 import { TEXT_CAR_POSITION } from './dom/const.js';
+import { VALIDE_CONDITION_RACING_COUNT } from './game/const.js';
 
 export default class RacingGameManager {
   constructor() {
     this.cars = [];
     this.racingCount = 0;
     this.winners = [];
+  }
+
+  static isValidCarNamesInput(input) {
+    const carNames = toCarNameList(input);
+
+    return (
+      !hasDuplicate(carNames) &&
+      carNames.every((carName) => isValidCarName(carName))
+    );
+  }
+
+  static isValidRacingCountInput(input) {
+    const count = Number(input);
+    const { min } = VALIDE_CONDITION_RACING_COUNT;
+
+    return Number.isInteger(count) && count >= min;
   }
 
   getCars() {
